requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
python-decouple==3.8
psycopg2==2.9.10; sys_platform == 'windows'
psycopg2-binary==2.9.10; sys_platform == 'linux' or sys_platform == "darwin"